Introduce 'logging' framework.
It seems to me that most applications MUST NOT call module-level
methods (e.g. logging.info) and SHOULD create its own logger.
Yes, it's very awkward to have a global variable 'log', but it is
much better to call getLogger() in every class and function (note
that the example shown in PEP 282 also does this way).