Make `fetch_copy()` to use the new state file.
This speeds up message retrieval noticeably. Below is a casual
performance measurement using a local POP3 server (which contains
2028 messages):

Before:
    $ /usr/bin/time -p python3 -m eubin.main -q test@example.conf
    real 0.69
    user 0.49
    sys 0.06

After:
    $ /usr/bin/time -p python3 -m eubin.main -q test@example.conf
    real 0.20
    user 0.12
    sys 0.01

In real usage (i.e. communicating with a remote host), the difference
should be much more significant because the new logic does not download
the header lines of messages (which has been producing a fair amount of
network traffic in the old implementation).