Clean up the main script and spool up the utility functions.
This should change no behaviour at all.

The old fetch_new_mail() function was a bit tangled because it tried
to do two things in parallel: 1) parse values from configuration file
and 2) set up a POP3 client and run it to fetch messages.

This introduces load_config() to sort out the flow of execution. Also,
it moves the utility functions from util.py to main.py, since these
funcitons are not in fact used elsewhere.